refactor: moved logic to seperated classes
